backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyJWT[crypto]==2.10.1
requests==2.32.3
cryptography==43.0.1
Pillow==10.4.0
reportlab==4.2.2
qrcode[pil]==7.4.2
boto3==1.34.144
